pyproject.toml
[tool.poetry]
name = "dddpy"
version = "0.1.0"
description = "A framework to support ddd python projects"
authors = ["Yuichiro Smith <contact@yu-smith.com>"]
license = "Apache License 2.0"
readme = "README.md"
repository = "https://github.com/aeroworks-io/dddpy"
keywords = ["DDD", "Domain"]
classifiers = [
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/aeroworks-io/dddpy/issues"
"Coverage Report" = "https://codecov.io/gh/aeroworks-io/dddpy/branch/main"

[tool.poetry.dependencies]
python = "^3.8"
pydantic = "^1.7.2"
Inject = "^4.3.1"
Faker = ">= 4.0.0, < 7.0.0"
python-ulid = "^1.0.2"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pytest = "^6.1.2"
coverage = "^5.3"
pytest-cov = "^2.10.1"
codecov = "^2.1.10"

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
python_classes = [
    "Test*",
    "*Test",
]
python_files = [
    "test_*.py",
    "*_test.py",
    "__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "@abstract"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
